refactor: change naming of fakes to fakes
