+++ packages/app-runtime/test/lib/natives/FakeNativeBootstrapper.ts
@@ -1,12 +1,12 @@
 import { EventEmitter2EventBus, Result } from "@js-soft/ts-utils";
 import { WebLoggerFactory } from "@js-soft/web-logger";
-import { INativeBootstrapper, INativeEnvironment } from "../../src";
-import { NativeConfigAccessMock } from "../mocks/NativeConfigAccessMock";
-import { NativeDatabaseFactoryMock } from "../mocks/NativeDatabaseFactoryMock";
-import { NativeDeviceInfoAccessMock } from "../mocks/NativeDeviceInfoAccessMock";
-import { NativeNotificationAccessMock } from "../mocks/NativeNotificationAccessMock";
+import { INativeBootstrapper, INativeEnvironment } from "../../../src";
+import { FakeNativeConfigAccess } from "./FakeNativeConfigAccess";
+import { FakeNativeDatabaseFactory } from "./FakeNativeDatabaseFactory";
+import { FakeNativeDeviceInfoAccess } from "./FakeNativeDeviceInfoAccess";
+import { FakeNativeNotificationAccess } from "./FakeNativeNotificationAccess";
 
-export class NativeBootstrapperMock implements INativeBootstrapper {
+export class FakeNativeBootstrapper implements INativeBootstrapper {
     private _nativeEnvironment: INativeEnvironment;
     public get nativeEnvironment(): INativeEnvironment {
         return this._nativeEnvironment;
@@ -21,14 +21,14 @@ export class NativeBootstrapperMock implements INativeBootstrapper {
         const nativeLogger = loggerFactory.getLogger("NativeMocks");
 
         this._nativeEnvironment = {
-            configAccess: new NativeConfigAccessMock(),
-            databaseFactory: new NativeDatabaseFactoryMock(),
-            deviceInfoAccess: new NativeDeviceInfoAccessMock(),
+            configAccess: new FakeNativeConfigAccess(),
+            databaseFactory: new FakeNativeDatabaseFactory(),
+            deviceInfoAccess: new FakeNativeDeviceInfoAccess(),
             eventBus: new EventEmitter2EventBus(() => {
                 // noop
             }),
             loggerFactory,
-            notificationAccess: new NativeNotificationAccessMock(nativeLogger)
+            notificationAccess: new FakeNativeNotificationAccess(nativeLogger)
         };
 
         await this._nativeEnvironment.deviceInfoAccess.init();
