+++ packages/app-runtime/test/lib/natives/FakeNativeDatabaseFactory.ts
@@ -1,7 +1,7 @@
 import { ILokiJsDatabaseFactory } from "@js-soft/docdb-access-loki";
 import loki from "lokijs";
 
-export class NativeDatabaseFactoryMock implements ILokiJsDatabaseFactory {
+export class FakeNativeDatabaseFactory implements ILokiJsDatabaseFactory {
     public create(name: string, options?: Partial<LokiConstructorOptions> & Partial<LokiConfigOptions> & Partial<ThrottledSaveDrainOptions>): Loki {
         return new loki(name, { ...options, persistenceMethod: "memory" });
     }
