+++ packages/app-runtime/test/lib/natives/FakeNativeDeviceInfoAccess.ts
@@ -1,7 +1,7 @@
 import { Result } from "@js-soft/ts-utils";
-import { INativeDeviceInfo, INativeDeviceInfoAccess } from "../../src";
+import { INativeDeviceInfo, INativeDeviceInfoAccess } from "../../../src";
 
-export class NativeDeviceInfoAccessMock implements INativeDeviceInfoAccess {
+export class FakeNativeDeviceInfoAccess implements INativeDeviceInfoAccess {
     public get deviceInfo(): INativeDeviceInfo {
         return {
             uuid: "00000000-7e7a-4e82-bd56-9cdba102ac13",
